--- projects/PTSD-UI/src/UIImpl.cpp
@@ -1,15 +1,0 @@
-#include "UIImpl.h"
-
-int PTSD::UIImpl::Init()
-{
-	return 0;
-}
-
-bool PTSD::UIImpl::Render()
-{
-	return false;
-}
-
-void PTSD::UIImpl::Shutdown()
-{
-}
